git-svn-id: file:///svn/toku/tokudb.1032b@7814 c7de825b-a66e-492c-adef-691d508d4ae1

+++ src/tdbtrace.c
@@ -267,7 +267,7 @@ static int tokutrace_db_open(DB * db, DB_TXN * txn, const char *fname, const cha
     if (tracefile) {
 	fprintf(tracefile, "db_open %d %d %d %s %s %d %u %d\n",
 		r,
-		dbpairnum(db), txnpairnum(txn), fname, dbname, dbtype, flags, mode);
+		dbpairnum(db), txnpairnum(txn), fname, dbname, (int)dbtype, flags, mode);
     }
     tunlock();
     return r;
